New Room and Cat
Added new room, invisible walls with low walls to represent them, and the cat.

The cat is using an internet 3d model temporarily, it moves towards the closest food, meowing and growing large upon eating each one.

The cat has vision and hearing however that is not in use yet.