BoxPlaceOnTopOfCat
Place boxes/cubes mostly on top of cat, disappearing when the cat continues moving.

Currently there is a bug where you can pick up this item even after placing it on the cat; will be fixed for the alpha as a result of changing pickup/item use system, so it should no longer be a problem then.